+++ CHANGELOG.md
@@ -7,8 +7,10 @@
 ### Changed
 
 - Ensure Python control can be found [#63][63]
+- Ensure activity is looked up before control is applied [#64][64]
 
 [63]: https://github.com/chaostoolkit/chaostoolkit-lib/issues/63
+[64]: https://github.com/chaostoolkit/chaostoolkit-lib/issues/64
 
 ## [1.0.0rc1][] - 2018-11-30