+++ CHANGELOG.md
@@ -9,10 +9,13 @@
 - Ensure Python control can be found [#63][63]
 - Ensure activity is looked up before control is applied [#64][64]
 - Ensure controls are applied only once before/after [#65][65]
+- Declare the name property in the catch block of the python control
+  validator [#66][66]
 
 [63]: https://github.com/chaostoolkit/chaostoolkit-lib/issues/63
 [64]: https://github.com/chaostoolkit/chaostoolkit-lib/issues/64
 [65]: https://github.com/chaostoolkit/chaostoolkit-lib/issues/65
+[66]: https://github.com/chaostoolkit/chaostoolkit-lib/issues/66
 
 ## [1.0.0rc1][] - 2018-11-30